requirements.txt
pyotp
pytelegrambotapi
python-dotenv
